linkage.c: Avoid conversion warning with word char position

+++ link-grammar/linkage/linkage.c
@@ -858,7 +858,7 @@ size_t linkage_get_word_byte_end(const Linkage linkage, WordIdx w)
 size_t linkage_get_word_char_start(const Linkage linkage, WordIdx w)
 {
 	if (linkage->num_words <= w) return 0; /* bounds-check */
-	int pos = linkage->wg_path_display[w]->start - linkage->sent->orig_sentence;
+	int pos = (int)(linkage->wg_path_display[w]->start - linkage->sent->orig_sentence);
 	char *sentchunk = alloca(pos+1);
 	strncpy(sentchunk, linkage->sent->orig_sentence, pos);
 	sentchunk[pos] = '\0';
@@ -868,7 +868,7 @@ size_t linkage_get_word_char_start(const Linkage linkage, WordIdx w)
 size_t linkage_get_word_char_end(const Linkage linkage, WordIdx w)
 {
 	if (linkage->num_words <= w) return 0; /* bounds-check */
-	int pos = linkage->wg_path_display[w]->end - linkage->sent->orig_sentence;
+	int pos = (int)(linkage->wg_path_display[w]->end - linkage->sent->orig_sentence);
 	char *sentchunk = alloca(pos+1);
 	strncpy(sentchunk, linkage->sent->orig_sentence, pos);
 	sentchunk[pos] = '\0';
